+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.9.16 - Unreleased
 
+### Added
+
+* Add "regular" and "star" classmethods to Polygon
+
 
 ## Version 0.9.15 - Released 2023-04-04